requirements.txt
wikipedia-api
gTTS
pexels-api
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
mutagen